warn: sửa warning lặt vặt

+++ src/game/model/entity/MovableObject.java
@@ -19,11 +19,11 @@ public abstract class MovableObject extends GameObject {
     }
 
     public void setDx(int dx) {
-        this.dx = (double) dx;
+        this.dx = dx;
     }
 
     public void setDy(int dy) {
-        this.dy = (double) dy;
+        this.dy = dy;
     }
 
     public void reverseDx() {
